+++ packages/core/src/lib/index.ts
@@ -10,7 +10,7 @@ export { default as DisposableObject } from './internal/DisposableObject.svelte'
 
 // components (v6)
 export { T, extend } from './components/T/T'
-export type { Props, Events, Slots } from './components/T/types'
+export type { Props } from './components/T/types'
 
 // plugins
 export { injectPlugin } from './plugins/injectPlugin'
@@ -20,16 +20,11 @@ export type { Plugin, NamedPlugin, PluginProps } from './plugins/types'
 // hooks
 export { useTask, type ThrelteUseTask, type ThrelteUseTaskOptions } from './hooks/useTask'
 export { useStage } from './hooks/useStage'
-export { useThrelte } from './hooks/useThrelte'
+export { useThrelte } from './context/compounds/useThrelte'
 export { useThrelteUserContext } from './hooks/useThrelteUserContext'
 
 // task scheduling system types
 export type { Key, Stage, Task, Scheduler, Schedule, TaskCallback } from './frame-scheduling'
-
-// legacy hooks
-// TODO: remove in Threlte 7
-export { useFrame, type ThrelteUseFrameOptions } from './hooks/legacy/useFrame'
-export { useRender } from './hooks/legacy/useRender'
 
 // useLoader
 export {
@@ -46,7 +41,8 @@ export { useParent } from './hooks/useParent'
 export { type AsyncWritable, asyncWritable } from './lib/asyncWritable'
 
 // contexts
-export { type ThrelteContext, createThrelteContext } from './lib/contexts'
+export { type ThrelteContext } from './context/compounds/useThrelte'
+export { createThrelteContext } from './context/createThrelteContext.svelte'
 
 // types
 export type { Size } from './types'
@@ -54,7 +50,7 @@ export type { Size } from './types'
 // utils
 export { createObjectStore } from './lib/createObjectStore'
 export { watch, memoize, type CurrentWritable, currentWritable } from './lib/storeUtils'
-export { useCache } from './lib/cache'
+export { useCache, createCacheContext } from './context/fragments/cache'
 export { resolvePropertyPath } from './components/T/utils/resolvePropertyPath'
 export { revision } from './lib/revision'
 
@@ -68,16 +64,21 @@ export type { SceneGraphObjectProperties } from './internal/SceneGraphObject.sve
 /**
  * @deprecated createRawEventDispatcher() has been removed from Threlte and is not compatible with Svelte 5.
  *
- * To send events, you must migrate your component to runes mode and use the rest props $$events object:
+ * To send events, you must migrate your component to runes mode and use callback props instead of events:
  *
  * @example
  * ```html
+ * <!-- MyComponent.svelte -->
  * <script>
- *   let { ...props } = $props()
+ *   let { onchange } = $props()
  *
  *   // at some point...
- *   props.$$events.change?.()
+ *   onchange?.()
  * </script>
+ *
+ * <!-- MyOtherComponent.svelte -->
+ * <MyComponent onchange={() => console.log('changed')} />
+ *
  * ```
  */
 export const createRawEventDispatcher = () => {