+++ packages/create-threlte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-threlte
 
+## 2.0.2-next.0
+
+### Patch Changes
+
+- f7c501d: Remove legacy useFrame and useRender hooks
+
+## 2.0.2
+
+### Patch Changes
+
+- 99101b4: Remove default vite configs
+
 ## 2.0.1
 
 ### Patch Changes