+++ packages/flex/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @threlte/flex
 
+## 2.0.0-next.3
+
+### Patch Changes
+
+- 36128d5: Replace events with callback props and slots with snippets
+
+## 2.0.0-next.2
+
+### Patch Changes
+
+- f7c501d: Remove legacy useFrame and useRender hooks
+
 ## 2.0.0-next.1
 
 ### Patch Changes